move XXX_with_retry into bza interface
