+++ bzt/modules/blazemeter/net_utils.py
@@ -16,10 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 import re
-import time
-import traceback
 from collections import namedtuple
-from functools import wraps
 from ssl import SSLError
 from urllib.error import URLError
 
@@ -28,39 +25,6 @@ from requests.exceptions import ReadTimeout
 from bzt import TaurusNetworkError
 
 NETWORK_PROBLEMS = (IOError, URLError, SSLError, ReadTimeout, TaurusNetworkError)
-
-
-def send_with_retry(method):
-    @wraps(method)
-    def _impl(self, *args, **kwargs):
-        try:
-            method(self, *args, **kwargs)
-        except (IOError, TaurusNetworkError):
-            self.log.debug("Error sending data: %s", traceback.format_exc())
-            self.log.warning("Failed to send data, will retry in %s sec...", self._user.timeout)
-            try:
-                time.sleep(self._user.timeout)
-                method(self, *args, **kwargs)
-                self.log.info("Succeeded with retry")
-            except NETWORK_PROBLEMS:
-                self.log.error("Fatal error sending data: %s", traceback.format_exc())
-                self.log.warning("Will skip failed data and continue running")
-
-    return _impl
-
-
-def get_with_retry(method):
-    @wraps(method)
-    def _impl(self, *args, **kwargs):
-        while True:
-            try:
-                return method(self, *args, **kwargs)
-            except NETWORK_PROBLEMS:
-                self.log.debug("Error making request: %s", traceback.format_exc())
-                self.log.warning("Failed to make request, will retry in %s sec...", self.user.timeout)
-                time.sleep(self.user.timeout)
-
-    return _impl
 
 
 def parse_blazemeter_test_link(link):
